requirements.txt
requests
selenium
BeautifulSoup4
pandas
jupyter
requests
python-dotenv
